Verilog implementation: simulation and synthesis
Added the verilog implementation of Li's (modified for the erroneous
ones) routines.
The verilate script generates all instances for bit sizes between 3 and 32, and simulate
them using verilator, checking with the usual division operation.
The synthesize script uses vivado to generate the FPGA implementation of the
circuit for the same bit sizes.